refactor: merge password-reset-token into authentication
