+++ src/apps/backend/modules/authentication/authentication_service.py
@@ -1,7 +1,17 @@
 from datetime import datetime, timedelta
 
 import jwt
+import urllib.parse
 
+from modules.account.errors import AccountBadRequestError
+from modules.account.internal.account_reader import AccountReader
+from modules.communication.email_service import EmailService
+from modules.communication.types import EmailRecipient, EmailSender, SendEmailParams
+from modules.config.config_service import ConfigService
+from modules.authentication.internals.password_reset_token.password_reset_token_reader import PasswordResetTokenReader
+from modules.authentication.internals.password_reset_token.password_reset_token_util import PasswordResetTokenUtil
+from modules.authentication.internals.password_reset_token.password_reset_token_writer import PasswordResetTokenWriter
+from modules.authentication.types import CreatePasswordResetTokenParams, PasswordResetToken
 from modules.authentication.errors import AccessTokenExpiredError, AccessTokenInvalidError
 from modules.authentication.types import (
     AccessToken,
@@ -9,7 +19,6 @@ from modules.authentication.types import (
     EmailBasedAuthAccessTokenRequestParams,
     OTPBasedAuthAccessTokenRequestParams,
 )
-from modules.account.account_service import AccountService
 from modules.account.internal.account_reader import AccountReader
 from modules.account.types import Account, AccountSearchParams
 from modules.config.config_service import ConfigService
@@ -29,7 +38,7 @@ class AuthenticationService:
 
     @staticmethod
     def create_access_token_by_phone_number(*, params: OTPBasedAuthAccessTokenRequestParams) -> AccessToken:
-        account = AccountService.get_account_by_phone_number(phone_number=params.phone_number)
+        account = AccountReader.get_account_by_phone_number(phone_number=params.phone_number)
 
         otp = OTPService.verify_otp(params=VerifyOTPParams(phone_number=params.phone_number, otp_code=params.otp_code))
 
@@ -62,3 +71,68 @@ class AuthenticationService:
             raise AccessTokenExpiredError(message="Access token has expired. Please login again.")
 
         return AccessTokenPayload(account_id=verified_token.get("account_id"))
+
+    @staticmethod
+    def create_password_reset_token(params: CreatePasswordResetTokenParams) -> PasswordResetToken:
+        account_obj = AccountReader.get_account_by_username(username=params.username)
+        token = PasswordResetTokenUtil.generate_password_reset_token()
+        password_reset_token = PasswordResetTokenWriter.create_password_reset_token(account_obj.id, token)
+        AuthenticationService.send_password_reset_email(
+            account_obj.id, account_obj.first_name, account_obj.username, token
+        )
+
+        return password_reset_token
+
+    @staticmethod
+    def get_password_reset_token_by_account_id(account_id: str) -> PasswordResetToken:
+        return PasswordResetTokenReader.get_password_reset_token_by_account_id(account_id)
+
+    @staticmethod
+    def set_password_reset_token_as_used_by_id(password_reset_token_id: str) -> PasswordResetToken:
+        return PasswordResetTokenWriter.set_password_reset_token_as_used(password_reset_token_id)
+
+    @staticmethod
+    def verify_password_reset_token(account_id: str, token: str) -> PasswordResetToken:
+        password_reset_token = AuthenticationService.get_password_reset_token_by_account_id(account_id)
+
+        if password_reset_token.is_expired:
+            raise AccountBadRequestError(
+                f"Password reset link is expired for accountId {account_id}. Please retry with new link"
+            )
+        if password_reset_token.is_used:
+            raise AccountBadRequestError(
+                f"Password reset is already used for accountId {account_id}. Please retry with new link"
+            )
+
+        is_token_valid = PasswordResetTokenUtil.compare_password(
+            password=token, hashed_password=password_reset_token.token
+        )
+        if not is_token_valid:
+            raise AccountBadRequestError(
+                f"Password reset link is invalid for accountId {account_id}. Please retry with new link."
+            )
+
+        return password_reset_token
+
+    @staticmethod
+    def send_password_reset_email(account_id: str, first_name: str, username: str, password_reset_token: str) -> None:
+
+        web_app_host = ConfigService[str].get_value(key="web_app_host")
+        default_email = ConfigService[str].get_value(key="mailer.default_email")
+        default_email_name = ConfigService[str].get_value(key="mailer.default_email_name")
+        forgot_password_mail_template_id = ConfigService[str].get_value(key="mailer.forgot_password_mail_template_id")
+
+        template_data = {
+            "first_name": first_name,
+            "password_reset_link": f"{web_app_host}/accounts/{account_id}/reset_password?token={urllib.parse.quote(password_reset_token)}",
+            "username": username,
+        }
+
+        password_reset_email_params = SendEmailParams(
+            template_id=forgot_password_mail_template_id,
+            recipient=EmailRecipient(email=username),
+            sender=EmailSender(email=default_email, name=default_email_name),
+            template_data=template_data,
+        )
+
+        EmailService.send_email(params=password_reset_email_params)
