+++ src/apps/backend/modules/authentication/errors.py
@@ -1,5 +1,6 @@
 from modules.authentication.types import AccessTokenErrorCode
 from modules.error.custom_errors import AppError
+from modules.authentication.types import PasswordResetTokenErrorCode
 
 
 class AccessTokenInvalidError(AppError):
@@ -27,3 +28,16 @@ class AuthorizationHeaderNotFoundError(AppError):
 class InvalidAuthorizationHeaderError(AppError):
     def __init__(self, message: str) -> None:
         super().__init__(code=AccessTokenErrorCode.INVALID_AUTHORIZATION_HEADER, http_status_code=401, message=message)
+
+from modules.error.custom_errors import AppError
+from modules.authentication.types import PasswordResetTokenErrorCode
+
+
+class PasswordResetTokenNotFoundError(AppError):
+
+    def __init__(self) -> None:
+        super().__init__(
+            code=PasswordResetTokenErrorCode.PASSWORD_RESET_TOKEN_NOT_FOUND,
+            http_status_code=404,
+            message=f"System is unable to find a token with this account",
+        )
