+++ src/apps/backend/modules/authentication/types.py
@@ -38,3 +38,25 @@ class AccessTokenErrorCode:
     AUTHORIZATION_HEADER_NOT_FOUND: str = "ACCESS_TOKEN_ERR_03"
     INVALID_AUTHORIZATION_HEADER: str = "ACCESS_TOKEN_ERR_04"
     ACCESS_TOKEN_INVALID: str = "ACCESS_TOKEN_ERR_05"
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class PasswordResetToken:
+    id: str
+    account: str
+    expires_at: str
+    is_expired: bool
+    is_used: bool
+    token: str
+
+
+@dataclass(frozen=True)
+class CreatePasswordResetTokenParams:
+    username: str
+
+
+@dataclass(frozen=True)
+class PasswordResetTokenErrorCode:
+    PASSWORD_RESET_TOKEN_NOT_FOUND: str = "PASSWORD_RESET_TOKEN_ERR_01"
