+++ src/apps/backend/server.py
@@ -14,7 +14,6 @@ from modules.config.config_service import ConfigService
 from modules.error.custom_errors import AppError
 from modules.logger.logger import Logger
 from modules.logger.logger_manager import LoggerManager
-from modules.password_reset_token.rest_api.password_reset_token_rest_api_server import PasswordResetTokenRestApiServer
 
 load_dotenv()
 
@@ -43,13 +42,9 @@ if ConfigService.has_value("is_server_running_behind_proxy") and ConfigService[b
 ):
     app.wsgi_app = ProxyFix(app.wsgi_app)  # type: ignore
 
-# Register access token apis
-access_token_blueprint = AuthenticationRestApiServer.create()
-api_blueprint.register_blueprint(access_token_blueprint)
-
-# Register password reset token apis
-password_reset_token_blueprint = PasswordResetTokenRestApiServer.create()
-api_blueprint.register_blueprint(password_reset_token_blueprint)
+# Register authentication apis
+authentication_blueprint = AuthenticationRestApiServer.create()
+api_blueprint.register_blueprint(authentication_blueprint)
 
 # Register accounts apis
 account_blueprint = AccountRestApiServer.create()
